requirements.txt
glfw==1.10.1
numpy==1.18.1
Pillow==7.0.0
PyGLM==1.1.7
PyOpenGL==3.1.5
